software-map: move node measurement scheduling to c4-node-measurement
Agent-Session: 01a067ce-34b7-7952-a023-ced44207d9ca
